+++ cli/src/commands/commit.rs
@@ -103,6 +103,9 @@ pub(crate) async fn cmd_commit(
     let diff_selector =
         workspace_command.diff_selector(ui, args.tool.as_deref(), args.interactive)?;
     let text_editor = workspace_command.text_editor()?;
+    let derive_tracked_from_ignores = workspace_command
+        .settings()
+        .get_bool("snapshot.derive-tracked-from-ignores")?;
     let mut tx = workspace_command.start_transaction();
     let base_tree = commit.parent_tree(tx.repo()).await?;
     let format_instructions = || {
@@ -127,6 +130,7 @@ new working-copy commit.
             ),
             matcher.as_ref(),
             format_instructions,
+            derive_tracked_from_ignores,
         )
         .await?;
     if !args.paths.is_empty() && tree.tree_ids() == base_tree.tree_ids() {