+++ lib/testutils/src/lib.rs
@@ -158,6 +158,7 @@ pub fn empty_snapshot_options() -> SnapshotOptions<'static> {
         start_tracking_matcher: &EverythingMatcher,
         force_tracking_matcher: &NothingMatcher,
         max_new_file_size: u64::MAX,
+        derive_tracked_from_ignores: false,
     }
 }