+++ CHANGELOG.md
@@ -426,6 +426,9 @@ None
   revisions by default (defined by `revsets.op-diff-changes-in`). A new flag,
   `--show-changes-in`, can be used to override this. [#6083](https://github.com/jj-vcs/jj/issues/6083)
 
+* New `snapshot.derive-tracked-from-ignores` option to treat ignoring/unignoring files
+  as deletes/adds, respectively.
+
 ### Fixed bugs
 
 * `.gitignore` with UTF-8 BOM can now be parsed correctly.